+++ src/main/kotlin/dev/mizarc/bellclaims/utils/ClaimFlagDisplay.kt
@@ -18,6 +18,8 @@ fun Flag.getIcon(): ItemStack {
         Flag.MobGriefing -> ItemStack(Material.CREEPER_HEAD)
         Flag.Pistons -> ItemStack(Material.PISTON)
         Flag.Fluids -> ItemStack(Material.WATER_BUCKET)
+        Flag.Trees -> ItemStack(Material.OAK_SAPLING)
+        Flag.Sculk -> ItemStack(Material.SCULK_CATALYST)
     }
 }
 
@@ -33,6 +35,8 @@ fun Flag.getDisplayName(): String {
         Flag.MobGriefing -> getLangText("NameFlagMobGriefing")
         Flag.Pistons -> getLangText("NameFlagPistons")
         Flag.Fluids -> getLangText("NameFlagFluids")
+        Flag.Trees -> getLangText("NameFlagTrees")
+        Flag.Sculk -> getLangText("NameFlagSculk")
     }
 }
 
@@ -48,5 +52,7 @@ fun Flag.getDescription(): String {
         Flag.MobGriefing -> getLangText("DescFlagMobGriefing")
         Flag.Pistons -> getLangText("DescFlagPistons")
         Flag.Fluids -> getLangText("DescFlagFluids")
+        Flag.Trees -> getLangText("DescFlagTrees")
+        Flag.Sculk -> getLangText("DescFlagSculk")
     }
 }